refactor: extract gift-wrapped messaging functions into gift-wrap.ts
