+++ components/messages/__tests__/messages.test.tsx
@@ -10,6 +10,7 @@ import Messages from "../messages";
 import { ChatsContext } from "../../../utils/context/context";
 import { SignerContext } from "@/components/utility-components/nostr-context-provider";
 import * as nostrHelper from "@/utils/nostr/nostr-helper-functions";
+import * as giftWrap from "@/utils/nostr/gift-wrap";
 import * as keypressHandler from "@/utils/keypress-handler";
 import { useRouter } from "next/router";
 
@@ -80,9 +81,11 @@ jest.mock("nostr-tools", () => ({
 }));
 
 jest.mock("@/utils/nostr/nostr-helper-functions");
+jest.mock("@/utils/nostr/gift-wrap");
 jest.mock("@/utils/keypress-handler");
 
 const mockNostrHelper = nostrHelper as jest.Mocked<typeof nostrHelper>;
+const mockGiftWrap = giftWrap as jest.Mocked<typeof giftWrap>;
 const mockUseKeyPress = keypressHandler.useKeyPress as jest.Mock;
 
 describe("Messages Component", () => {
@@ -264,16 +267,14 @@ describe("Messages Component", () => {
       created_at: 1005,
       tags: [["p", mockChatPubkey1]],
     };
-    mockNostrHelper.constructGiftWrappedEvent.mockResolvedValue(
+    mockGiftWrap.constructGiftWrappedEvent.mockResolvedValue(
       mockGiftWrappedEvent as any
     );
-    mockNostrHelper.constructMessageSeal.mockResolvedValue(
-      "sealed-event" as any
-    );
-    mockNostrHelper.constructMessageGiftWrap.mockResolvedValue(
+    mockGiftWrap.constructMessageSeal.mockResolvedValue("sealed-event" as any);
+    mockGiftWrap.constructMessageGiftWrap.mockResolvedValue(
       "gift-wrapped-event" as any
     );
-    mockNostrHelper.sendGiftWrappedMessageEvent.mockResolvedValue(undefined);
+    mockGiftWrap.sendGiftWrappedMessageEvent.mockResolvedValue(undefined);
 
     renderComponent();
     await waitFor(() => {
@@ -284,7 +285,7 @@ describe("Messages Component", () => {
       fireEvent.click(sendButton);
     });
     await waitFor(() => {
-      expect(mockNostrHelper.constructGiftWrappedEvent).toHaveBeenCalledWith(
+      expect(mockGiftWrap.constructGiftWrappedEvent).toHaveBeenCalledWith(
         mockUserPubkey,
         mockChatPubkey1,
         "Test message",
@@ -296,7 +297,7 @@ describe("Messages Component", () => {
   it("should handle errors when sending a message", async () => {
     mockChatsContextValue.isLoading = false;
     mockChatsContextValue.chatsMap = mockChatsMap;
-    mockNostrHelper.constructGiftWrappedEvent.mockRejectedValue(
+    mockGiftWrap.constructGiftWrappedEvent.mockRejectedValue(
       new Error("Send failed")
     );
