+++ components/product-invoice-card.tsx
@@ -50,15 +50,17 @@ import {
 } from "@/utils/cashu/wallet-recovery";
 import { persistBuyerP2pkEscrowRecord } from "@/utils/cashu/p2pk-escrow-records";
 import {
-  constructGiftWrappedEvent,
-  constructMessageSeal,
-  constructMessageGiftWrap,
-  sendGiftWrappedMessageEvent,
   getLocalStorageData,
   publishProofEvent,
   generateKeys,
   getSavedAddresses,
 } from "@/utils/nostr/nostr-helper-functions";
+import {
+  constructGiftWrappedEvent,
+  constructMessageSeal,
+  constructMessageGiftWrap,
+  sendGiftWrappedMessageEvent,
+} from "@/utils/nostr/gift-wrap";
 import { LightningAddress } from "@getalby/lightning-tools";
 import QRCode from "qrcode";
 import { v4 as uuidv4 } from "uuid";
