+++ utils/nostr/__tests__/nostr-helper-functions.test.ts
@@ -43,9 +43,6 @@ jest.mock("nostr-tools", () => {
 
 import {
   blossomUploadImages,
-  constructGiftWrappedEvent,
-  constructMessageGiftWrap,
-  constructMessageSeal,
   createBlossomServerEvent,
   createNostrDeleteEvent,
   createNostrProfileEvent,
@@ -76,13 +73,18 @@ import {
   publishReviewEvent,
   REPORT_TYPES,
   saveNWCString,
-  sendGiftWrappedMessageEvent,
   setLocalStorageDataOnSignIn,
   validateNPubKey,
   validateNSecKey,
   verifyNip05Identifier,
   withBlastr,
 } from "../nostr-helper-functions";
+import {
+  constructGiftWrappedEvent,
+  constructMessageGiftWrap,
+  constructMessageSeal,
+  sendGiftWrappedMessageEvent,
+} from "../gift-wrap";
 import {
   approveCommunityPost,
   createCommunityPost,
